requirements.txt
fastapi
uvicorn
openai
requests
pydantic
python-dotenv
azure-ai-projects
azure-identity